backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
PyMySQL==1.1.0
pymongo==4.5.0
xarray==2023.8.0
pandas==2.0.3
numpy==1.24.3
PyJWT==2.8.0
Werkzeug==2.3.7
sentence-transformers==2.2.2
faiss-cpu==1.7.4
netCDF4==1.6.4
python-dotenv==1.0.0
gunicorn==21.2.0
